source code/requirements.txt
torch>=2.0.0
transformers>=4.30.0
langdetect>=1.0.9
llama-cpp-python>=0.2.0
requests>=2.28.0
numpy>=1.21.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0
regex>=2022.0.0
tqdm>=4.64.0
urllib3>=1.26.0
